requirements_python3_12_11.txt
adb-cloud-connector==1.0.2
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
ansi2html==1.9.2
anyio==4.11.0
asgiref==3.10.0
asttokens==3.0.0
async-generator==1.10
attrs==25.3.0
Authlib==1.6.5
babel==2.17.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
boto3==1.40.41
botocore==1.40.45
bs4==0.0.2
build==1.3.0
cachetools==6.2.0
cerebras_cloud_sdk==1.50.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
chroma-hnswlib==0.7.6
chromadb==1.0.6
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
comm==0.2.3
courlan==1.3.2
cryptography==46.0.2
cyclopts==3.24.0
dash==2.9.3
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_auth==2.3.0
dash_daq==0.6.0
dateparser==1.2.2
debugpy==1.8.17
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.2
duckduckgo_search==6.4.2
durationpy==0.10
email-validator==2.3.0
et_xmlfile==2.0.0
exa-py==1.15.6
exceptiongroup==1.3.0
executing==2.2.1
Faker==18.13.0
fakeredis==2.31.3
fastapi==0.115.9
fastmcp==2.12.4
fastuuid==0.13.5
filelock==3.19.1
fire==0.7.1
Flask==3.1.0
flatbuffers==25.9.23
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-genai==1.41.0
googleapis-common-protos==1.70.0
groq==0.32.0
grpcio==1.75.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
halo==0.0.31
hf-xet==1.1.10
hpack==4.1.0
htmldate==1.9.3
httpcore==1.0.9
httplib2==0.31.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
json_repair==0.52.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
jupyter-dash==0.4.2
jupyter_client==8.6.3
jupyter_core==5.8.1
jusText==3.0.2
kubernetes==34.1.0
langroid==0.53.16
lazy-object-proxy==1.12.0
litellm==1.77.5
log-symbols==0.0.14
lxml==5.4.0
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mcp==1.16.0
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.6.4
narwhals==2.6.0
nest-asyncio==1.6.0
nltk==3.9.2
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.23.0
openai==1.109.1
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.58b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.2.3
parse==1.20.2
parso==0.8.5
pathable==0.4.4
pdf2image==1.17.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.1
pluggy==1.6.0
portalocker==3.2.0
posthog==6.7.6
prettytable==3.16.0
primp==0.15.0
prompt_toolkit==3.0.52
propcache==0.4.0
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
PyGithub==1.59.1
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.26.4
pymupdf4llm==0.0.27
PyNaCl==1.6.0
pyparsing==3.2.5
pyperclip==1.11.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.4.2
pytest-rerunfailures==15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-magic==0.4.27
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
qdrant-client==1.15.1
rank-bm25==0.2.2
RapidFuzz==3.14.1
redis==5.3.1
referencing==0.36.2
regex==2025.9.18
requests==2.32.3
requests-oauthlib==1.3.1
retrying==1.4.2
rfc3339-validator==0.1.4
rich==13.9.4
rich-rst==1.3.1
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
setuptools==80.9.0
setuptools-scm==9.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
spinners==0.0.24
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.45.3
supervisor==4.3.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
thefuzz==0.22.1
tiktoken==0.11.0
tld==0.13.1
tokenizers==0.22.1
tornado==6.5.2
tqdm==4.67.1
trafilatura==1.12.2
traitlets==5.14.3
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.14
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.1
wget==3.2
wrapt==1.17.3
yarl==1.21.0
zipp==3.23.0
